Switch CentOS to netinstall & fix build.
Thanks to @juliandunn for pointing out that the build does not work
properly as development tools were not being installed.

This switches to downloading the package groups but it means the URL
needs to be specified in the kickstart config so you can no longer share
a config between other CentOS versions like the Bento DRY stuff.